ft_transcendence/srcs/django/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
idna==3.7
PyJWT==2.8.0
pyotp==2.9.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.0
urllib3==2.2.2
psycopg2==2.9.9
uvicorn==0.30.1
gunicorn==22.0.0
django-cors-headers==4.4.0
cryptography==42.0.8
